adjust file/folder structure
