+++ UNUMCanvas/Classes/CanvasController+RotateGesture.swift
