+++ UNUMCanvas/Classes/CanvasController+TapGesture.swift
